Improve performance by reducing buffer copying
There are two specific improvements we can make:

1. In DNGBASE.__process__() we may be able to use the memoryview
underlying the buffer directly, instead of copying it out using
numpy.tobytes().

2. In DNG.write() we don't copy the pixel data to self.buf, as this
copy is unnecessary if we're only going to write the data somewhere
else straight after. Instead, we write it directly to the final
output.

These optimisations are enabled by passing a parameter "file" to
DNGBASE.convert(). The parameter should be a BinaryIO object, such
as a handle to an open file, or a BytesIO.

When the "file" parameter is omitted, the code should behave exactly
as it did previously.